update to use socketio
the kaspa wallet npm module offers the ability for listening to changes to wallet balance.

to get that to work with this app, we need to set up some sort of socket/websocket connection so that the app is always listening to the server changes that occur.

socketio was implemented for `wallet balance` to address this situation